backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pydantic>=2.7.3
pydantic-settings>=2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-dateutil==2.8.2
openai==1.10.0
anthropic>=0.40.0
boto3==1.34.34
websockets==12.0
# CrewAI and mem0 for agent-based strategy execution
# Using latest compatible versions
crewai>=0.86.0
crewai-tools>=0.17.0,<1.0.0
mem0ai>=0.1.29
# VectorBT and its dependencies (Python 3.12 compatible)
numpy>=1.26.0,<2.0.0
pandas>=2.0.0,<3.0.0
vectorbt==0.26.2
yfinance==0.2.36
plotly==5.18.0
kaleido==0.2.1
# Alternative data sources
pycoingecko==3.1.0
requests==2.32.3
